Implement Chalk::Config#reload.
Add the ability to reload a specified configuration file.

This is useful for applications that need to pick up configuration from
an external source (e.g. puppet) and want to be able to gracefully
reload.

Add an interface to reload a single file, instead of everything, since
supporting targetted reload seems much saner than potentially having all
of your runtime configuration change out from under you on disk, but I'm
open to discussion on this point.